+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # TelegramBotAPI changelog
 
+## 0.38.16
+
+* `Core`:
+    * `TelegramAPIUrlsKeeper` now have two new things: properties `webAppDataSecretKey` and fun `checkWebAppLink`
+
 ## 0.38.15
 
 * `Common`: